+++ commando/evaluation.py
@@ -1,10 +1,14 @@
-from commando import ComManDo
 import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
 import seaborn as sns
 from sklearn.decomposition import PCA
+from sklearn.feature_selection import r_regression
+from sklearn.metrics import silhouette_samples, silhouette_score
 import torch
+
+from .commando import ComManDo
+from .utilities import ensure_list, predict_nn
 
 
 def test_partial(
@@ -45,6 +49,7 @@ def test_partial(
 
 
 def generate_figure(
+    # Data
     cm_trained,
     cm_data,
     dataset,
@@ -52,10 +57,22 @@ def generate_figure(
     alg_results=[],
     alg_names=[],
     dataset_names=None,
+    # Style
+    scale=20,
+    size_bound=(4, .25),
+    # Visualizations
+    reconstruction_features={},
+    # Remove Visualizations
+    exclude_predict=[],
+    skip_partial=False,
+    skip_nn=False,
 ):
     """Compares ComManDo with ``alg_results``"""
     # asddf: Legends may be wrong unless sorted
     assert len(alg_results) == len(alg_names), '``alg_*`` params must correspond.'
+
+    # Style
+    sns.set(style='darkgrid')
 
     # Setup
     integrated_data = [cm_data, *alg_results]
@@ -64,30 +81,50 @@ def generate_figure(
     num_modalities = len(dataset)
     num_algorithms = len(integrated_data)
 
-    fig = plt.figure(constrained_layout=True, figsize=(12, 16))
-    subfigs = fig.subfigures(5, 1, wspace=.07)
+    height_numerators = [
+        1,
+        1,
+        1,
+        1,
+        (num_modalities**2 - num_modalities) - len(exclude_predict)
+    ]
+    height_denominators = [
+        (num_modalities),
+        (num_modalities * num_algorithms),
+        (2 - skip_partial),
+        (num_modalities * num_algorithms),
+        (4 if not skip_nn else 3),
+    ]
+    height_denominators = [max(size_bound[0], x) for x in height_denominators]
+    height_ratios = [n/d for n, d in zip(height_numerators, height_denominators)]
+    height_ratios = [min(size_bound[1] * m, x) for m, x in zip(height_numerators, height_ratios)]
+    figsize = (scale, scale * sum(height_ratios))
+    fig = plt.figure(constrained_layout=True, figsize=figsize)
+    subfigs = fig.subfigures(5, 1, height_ratios=height_ratios, wspace=.07)
 
     # Raw Data
     cfig = subfigs[0]
-    axs = cfig.subplots(1, num_modalities)
-    for ax, i in zip(axs, range(num_modalities)):
+    for i in range(num_modalities):
+        ax = cfig.add_subplot(1, num_modalities, i+1)  # , projection='3d'
         pca_data = PCA(n_components=2).fit_transform(dataset[i])
         for label in np.unique(np.concatenate(labels)):
             pca_data_subset = np.transpose(pca_data[labels[i] == label])
             ax.scatter(*pca_data_subset, s=5., label=label)
+            ax.set_aspect('equal', adjustable='box')
         title = dataset_names[i] if dataset_names is not None else f'Dataset {i}'
         ax.set_title(title)
         ax.set_xlabel('PCA-1')
         ax.set_ylabel('PCA-2')
+        # ax.set_zlabel('PCA-3')
     ax.legend()
-    cfig.suptitle('Raw Data')
+    fig.suptitle('Raw Data')
 
     # Integrated Data
     cfig = subfigs[1]
     csubfigs = cfig.subfigures(1, num_algorithms, wspace=.07)
     for csubfig, j in zip(csubfigs, range(num_algorithms)):
-        axs = csubfig.subplots(1, num_modalities)
-        for ax, i in zip(axs, range(num_modalities)):
+        for i in range(num_modalities):
+            ax = csubfig.add_subplot(1, num_modalities, i+1)  # , projection='3d'
             for label in np.unique(np.concatenate(labels)):
                 data_subset = np.transpose(integrated_data[j][i][labels[i] == label])[:2, :]
                 ax.scatter(*data_subset, s=5., label=label)
@@ -95,32 +132,104 @@ def generate_figure(
             ax.set_title(title)
             ax.set_xlabel('Latent Feature 1')
             ax.set_ylabel('Latent Feature 2')
+            # ax.set_zlabel('Latent Feature 3')
         csubfig.suptitle(integrated_alg_names[j])
     cfig.suptitle('Integrated Embeddings')
 
-    # Distance by Cell Type
+    # Accuracy by Partial
     cfig = subfigs[2]
-    axs = cfig.subplots(1, num_algorithms)
-    for ax, i in zip(axs, range(num_algorithms)):
-        lab, dat = cm_trained.test_label_dist(integrated_data[i], labels, verbose=False)
-        # Sort to look nice
-        idx = np.argsort(dat, axis=1)[0]
-        dat = dat[idx, :][:, idx]
-        lab = lab[idx]
-        ax = sns.heatmap(dat, xticklabels=lab, yticklabels=lab, linewidth=0, cmap='YlGnBu', ax=ax)
-        ax.set_title(integrated_alg_names[i])
-    cfig.suptitle('Distance of Medoid by Cell Type')
+    csubfigs = ensure_list(
+        cfig.subfigures(1, 2 if not skip_partial else 1, wspace=.07))
+    csubfig_idx = 0
+    if not skip_partial:
+        ax = csubfigs[csubfig_idx].subplots(1, 1)
+        acc_list, fraction_range = test_partial(dataset, types, plot=False)
+        for k, v in acc_list.items():
+            ax.plot(fraction_range, v, '.-', label=k)
+        ax.set_title('Accuracy by Partial Alignment')
+        ax.set_xlabel('Fraction Assumed Aligned')
+        ax.set_ylabel('Statistic')
+        ax.legend()
+        csubfig_idx += 1
+    #
+    # Metric by Algorithm
+    ax = csubfigs[csubfig_idx].subplots(1, 1)
+    acc_dict = {
+        'Algorithm': integrated_alg_names,
+        'Label Transfer Accuracy': [],
+        'FOSCTTM': [],
+    }
+    for name in dataset_names:
+        acc_dict['Silhouette Score:\n' + name] = []
+    for i in range(num_algorithms):
+        acc_dict['Label Transfer Accuracy'].append(
+            cm_trained.test_LabelTA(integrated_data[i], types))
+        acc_dict['FOSCTTM'].append(
+            cm_trained.test_closer(integrated_data[i]))
+        for j, name in enumerate(dataset_names):
+            acc_dict['Silhouette Score:\n' + name].append(
+                silhouette_score(integrated_data[i][j], types[j]))
+    df = pd.DataFrame(acc_dict).melt(
+        id_vars=list(acc_dict.keys())[:1],
+        value_vars=list(acc_dict.keys())[1:])
+    sns.barplot(data=df, x='variable', y='value', hue='Algorithm', ax=ax, alpha=.6)
+    ax.set_xlabel(None)
+    ax.set_ylabel(None)
+    ax.set_title('Metric by Algorithm')
+    cfig.suptitle('Miscellaneous Accuracy Statistics')
+    csubfig_idx += 1
+
+    # Distance by Cell Type
+    # cfig = subfigs[3]
+    # axs = cfig.subplots(1, num_algorithms)
+    # for ax, i in zip(axs, range(num_algorithms)):
+    #     lab, dat = cm_trained.test_label_dist(integrated_data[i], labels, verbose=False)
+    #     # Sort to look nice
+    #     idx = np.argsort(dat, axis=1)[0]
+    #     dat = dat[idx, :][:, idx]
+    #     lab = lab[idx]
+    #     ax = sns.heatmap(dat, xticklabels=lab, yticklabels=lab, linewidth=0, cmap='YlGnBu', ax=ax)
+    #     ax.set_title(integrated_alg_names[i])
+    # cfig.suptitle('Distance of Medoid by Cell Type')
+
+    # Silhouette Value Boxplots
+    cfig = subfigs[3]
+    csubfigs = ensure_list(
+        cfig.subfigures(1, num_algorithms, wspace=.07))
+    for i in range(num_algorithms):
+        csubfig = csubfigs[i]
+        for j in range(num_modalities):
+            ax = csubfig.add_subplot(1, num_modalities, j+1)
+            # Calculate silhouette coefficients
+            coefs = silhouette_samples(integrated_data[i][j], types[j])
+            coefs_by_label = []
+            for label in np.unique(np.concatenate(labels)):
+                coefs_by_label.append(coefs[labels[j] == label])
+            ax.boxplot(coefs_by_label,
+                       labels=np.unique(np.concatenate(labels)),
+                       # notch=True,
+                       vert=True,
+                       patch_artist=True,
+                       boxprops={'facecolor': 'white'})
+            ax.set_xlabel('Cell')
+            ax.set_ylabel('Silhouette Coefficient')
+        csubfig.suptitle(integrated_alg_names[i])
+    cfig.suptitle('Silhouette Score by Cell Type')
 
     # Reconstruct Modality
-    cfig = subfigs[3]
-    csubfigs = cfig.subfigures(1, num_modalities**2 - num_modalities, wspace=.07)
+    cfig = subfigs[4]
+    csubfigs = ensure_list(cfig.subfigures(
+        (num_modalities**2 - num_modalities) - len(exclude_predict),
+        1,
+        wspace=.07
+    ))
     fig_idx = 0
     for i in range(num_modalities):
         for j in range(num_modalities):
-            if i == j:
+            if i == j or ((i, j) in exclude_predict):
                 continue
             csubfig = csubfigs[fig_idx]
-            axs = csubfig.subplots(1, 2)
+            axs = csubfig.subplots(1, 4 if not skip_nn else 3)
             fig_idx += 1
 
             if dataset_names is not None:
@@ -135,54 +244,52 @@ def generate_figure(
             actual = dataset[j]
 
             # asdf: Use PCA
+            # Setup
+            if (i, j) in reconstruction_features:
+                feat = reconstruction_features[(i, j)]
+            else:
+                feat = [0, 1]
+
+            axi = 0
             # Real
             for label in np.unique(np.concatenate(labels)):
-                subdata = np.transpose(actual[:, :2][labels[j] == label])
-                axs[0].scatter(*subdata, label=label, s=5.)
-            axs[0].set_title(f'Actual {dataset_names[j]}')
-            axs[0].set_xlabel('Latent Feature 1')
-            axs[0].set_ylabel('Latent Feature 2')
+                subdata = np.transpose(actual[:, feat][labels[j] == label])
+                axs[axi].scatter(*subdata, label=label, s=5.)
+            axs[axi].set_title(f'Actual {dataset_names[j]}')
+            axs[axi].set_xlabel('Latent Feature 1')
+            axs[axi].set_ylabel('Latent Feature 2')
+            axi += 1
 
             # Predicted
             for label in np.unique(np.concatenate(labels)):
-                subdata = np.transpose(predicted[:, :2][labels[j] == label])
-                axs[1].scatter(*subdata, label=label, s=5.)
-            axs[1].set_title('ComManDo Predicted')
-            axs[1].set_xlabel('Latent Feature 1')
-            axs[1].set_ylabel('Latent Feature 2')
+                subdata = np.transpose(predicted[:, feat][labels[j] == label])
+                axs[axi].scatter(*subdata, label=label, s=5.)
+            axs[axi].set_title('ComManDo Predicted')
+            axs[axi].set_xlabel('Latent Feature 1')
+            axs[axi].set_ylabel('Latent Feature 2')
+            axi += 1
+
+            # NN Predicted
+            if not skip_nn:
+                nn_predicted = predict_nn(
+                    torch.tensor(dataset[i]).float(), torch.tensor(dataset[j]).float())
+                for label in np.unique(np.concatenate(labels)):
+                    subdata = np.transpose(nn_predicted[:, feat][labels[j] == label])
+                    axs[axi].scatter(*subdata, label=label, s=5.)
+                axs[axi].set_title('NN Predicted')
+                axs[axi].set_xlabel('Latent Feature 1')
+                axs[axi].set_ylabel('Latent Feature 2')
+                axi += 1
+
+            # Correlation
+            corr = []
+            for k in range(predicted.shape[1]):
+                corr.append(r_regression(predicted[:, [k]], actual[:, k])[0])
+            axs[axi].bar(range(len(corr)), corr)
+            axs[axi].set_title('Correlation by Feature')
+            axs[axi].set_xlabel('Feature')
+            axs[axi].set_ylabel('Correlation')
+            axi += 1
     cfig.suptitle('Modality Prediction')
-
-    # Accuracy by Partial
-    cfig = subfigs[4]
-    csubfigs = cfig.subfigures(1, 2, wspace=.07)
-    ax = csubfigs[0].subplots(1, 1)
-    acc_list, fraction_range = test_partial(dataset, types, plot=False)
-    for k, v in acc_list.items():
-        ax.plot(fraction_range, v, '.-', label=k)
-    ax.set_title('Accuracy by Partial Alignment')
-    ax.set_xlabel('Fraction Assumed Aligned')
-    ax.set_ylabel('Statistic')
-    ax.legend()
-
-    ax = csubfigs[1].subplots(1, 1)
-    acc_dict = {
-        'Algorithm': integrated_alg_names,
-        'Label Transfer Accuracy': [],
-        'FOSCTTM': [],
-    }
-    for i in range(num_algorithms):
-        acc_dict['Label Transfer Accuracy'].append(
-            cm_trained.test_LabelTA(integrated_data[i], types))
-        acc_dict['FOSCTTM'].append(
-            cm_trained.test_closer(integrated_data[i]))
-    df = pd.DataFrame(acc_dict).melt(
-        id_vars=['Algorithm'],
-        value_vars=['Label Transfer Accuracy', 'FOSCTTM'])
-    sns.barplot(data=df, x='variable', y='value', hue='Algorithm', ax=ax,
-                palette='dark', alpha=.6)
-    ax.set_title('Metric by Algorithm')
-    ax.set_xlabel('Fraction Assumed Aligned')
-    ax.set_ylabel(None)
-    cfig.suptitle('Miscellaneous Accuracy Statistics')
 
     plt.show()